Redesigned game logic and interface
Create three buttons, one for each selection.
Add an event listener to the buttons that call
playRound function with the correct playerSelection
every time a button is clicked. Add a div for displaying results.
Display the running score, and announce a winner
of the game once one player reaches 5 points.
Refactor original code.